+++ packages/analytics/src/factory.ts
@@ -34,7 +34,7 @@ import {
   insertScriptTag,
   getOrCreateDataLayer,
   wrapOrCreateGtag,
-  hasDataLayer
+  findGtagScriptOnPage
 } from './helpers';
 import { ANALYTICS_ID_FIELD } from './constants';
 import { AnalyticsError, ERROR_FACTORY } from './errors';
@@ -107,11 +107,11 @@ export function factory(
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   extendApp: (props: { [prop: string]: any }) => void
 ): FirebaseAnalytics {
-  if (!app.options[ANALYTICS_ID_FIELD]) {
+  const analyticsId = app.options[ANALYTICS_ID_FIELD];
+  if (!analyticsId) {
     throw ERROR_FACTORY.create(AnalyticsError.NO_GA_ID);
   }
 
-  const analyticsId: string = app.options[ANALYTICS_ID_FIELD];
   if (initializedIdPromisesMap[analyticsId] != null) {
     throw ERROR_FACTORY.create(AnalyticsError.ALREADY_EXISTS, {
       id: analyticsId
@@ -122,9 +122,8 @@ export function factory(
     // Steps here should only be done once per page: creation or wrapping
     // of dataLayer and global gtag function.
 
-    // Presence of previously existing dataLayer used to detect if user has
-    // already put the gtag snippet on this page.
-    if (!hasDataLayer(dataLayerName)) {
+    // Detect if user has already put the gtag <script> tag on this page.
+    if (!findGtagScriptOnPage()) {
       insertScriptTag(dataLayerName);
     }
     getOrCreateDataLayer(dataLayerName);
@@ -159,7 +158,7 @@ export function factory(
     setUserProperties: (properties, options) =>
       setUserProperties(wrappedGtagFunction, analyticsId, properties, options),
     setAnalyticsCollectionEnabled: enabled =>
-      setAnalyticsCollectionEnabled(app.options[ANALYTICS_ID_FIELD], enabled)
+      setAnalyticsCollectionEnabled(analyticsId, enabled)
   };
 
   extendApp({