+++ packages/firestore/src/local/target_cache.ts
@@ -15,7 +15,6 @@
  * limitations under the License.
  */
 
-import { Query } from '../core/query';
 import { SnapshotVersion } from '../core/snapshot_version';
 import { ListenSequenceNumber, TargetId } from '../core/types';
 import { DocumentKeySet } from '../model/collections';
@@ -23,14 +22,16 @@ import { DocumentKey } from '../model/document_key';
 
 import { PersistenceTransaction } from './persistence';
 import { PersistencePromise } from './persistence_promise';
-import { QueryData } from './query_data';
+import { TargetData } from './target_data';
+import { Target } from '../core/target';
 
 /**
- * Represents cached queries received from the remote backend.
+ * Represents cached targets received from the remote backend.
  *
- * The cache is keyed by Query and entries in the cache are QueryData instances.
+ * The cache is keyed by `Target` and entries in the cache are `TargetData`
+ * instances.
  */
-export interface QueryCache {
+export interface TargetCache {
   /**
    * A global snapshot version representing the last consistent snapshot we
    * received from the backend. This is monotonically increasing and any
@@ -55,11 +56,11 @@ export interface QueryCache {
   ): PersistencePromise<ListenSequenceNumber>;
 
   /**
-   * Call provided function with each `QueryData` that we have cached.
+   * Call provided function with each `TargetData` that we have cached.
    */
   forEachTarget(
     txn: PersistenceTransaction,
-    f: (q: QueryData) => void
+    f: (q: TargetData) => void
   ): PersistencePromise<void>;
 
   /**
@@ -79,71 +80,71 @@ export interface QueryCache {
   /**
    * Adds an entry in the cache.
    *
-   * The cache key is extracted from `queryData.query`. The key must not already
+   * The cache key is extracted from `targetData.target`. The key must not already
    * exist in the cache.
    *
-   * @param queryData A QueryData instance to put in the cache.
+   * @param targetData A TargetData instance to put in the cache.
    */
-  addQueryData(
+  addTargetData(
     transaction: PersistenceTransaction,
-    queryData: QueryData
+    targetData: TargetData
   ): PersistencePromise<void>;
 
   /**
    * Updates an entry in the cache.
    *
-   * The cache key is extracted from `queryData.query`. The entry must already
+   * The cache key is extracted from `targetData.target`. The entry must already
    * exist in the cache, and it will be replaced.
-   * @param {QueryData} queryData The QueryData to be replaced into the cache.
+   * @param {TargetData} targetData The TargetData to be replaced into the cache.
    */
-  updateQueryData(
+  updateTargetData(
     transaction: PersistenceTransaction,
-    queryData: QueryData
+    targetData: TargetData
   ): PersistencePromise<void>;
 
   /**
-   * Removes the cached entry for the given query data. It is an error to remove
-   * a query data that does not exist.
+   * Removes the cached entry for the given target data. It is an error to remove
+   * a target data that does not exist.
    *
    * Multi-Tab Note: This operation should only be called by the primary client.
    */
-  removeQueryData(
+  removeTargetData(
     transaction: PersistenceTransaction,
-    queryData: QueryData
+    targetData: TargetData
   ): PersistencePromise<void>;
 
   /**
    * The number of targets currently in the cache.
    */
   // Visible for testing.
-  getQueryCount(
+  getTargetCount(
     transaction: PersistenceTransaction
   ): PersistencePromise<number>;
 
   /**
-   * Looks up a QueryData entry by query.
+   * Looks up a TargetData entry by target.
    *
-   * @param query The query corresponding to the entry to look up.
-   * @return The cached QueryData entry, or null if the cache has no entry for
-   * the query.
+   * @param target The query target corresponding to the entry to look up.
+   * @return The cached TargetData entry, or null if the cache has no entry for
+   * the target.
    */
-  getQueryData(
+  getTargetData(
     transaction: PersistenceTransaction,
-    query: Query
-  ): PersistencePromise<QueryData | null>;
+    target: Target
+  ): PersistencePromise<TargetData | null>;
 
   /**
-   * Looks up a QueryData entry by target ID.
+   * Looks up a TargetData entry by target ID.
    *
-   * @param targetId The target ID of the QueryData entry to look up.
-   * @return The cached QueryData entry, or null if the cache has no entry for
-   * the query.
+   * @param targetId The target ID of the TargetData entry to look up.
+   * @return The cached TargetData entry, or null if the cache has no entry for
+   * the target.
    */
   // PORTING NOTE: Multi-tab only.
-  getQueryDataForTarget(
+  getTargetDataForTarget(
     txn: PersistenceTransaction,
     targetId: TargetId
-  ): PersistencePromise<QueryData | null>;
+  ): PersistencePromise<TargetData | null>;
 
   /**
    * Adds the given document keys to cached query results of the given target