+++ packages/installations/src/functions/get-id.ts
@@ -19,7 +19,6 @@ import { FirebaseApp } from '@firebase/app-types';
 import { extractAppConfig } from '../helpers/extract-app-config';
 import { getInstallationEntry } from '../helpers/get-installation-entry';
 import { refreshAuthToken } from '../helpers/refresh-auth-token';
-import { RequestStatus } from '../interfaces/installation-entry';
 
 export async function getId(app: FirebaseApp): Promise<string> {
   const appConfig = extractAppConfig(app);
@@ -28,14 +27,11 @@ export async function getId(app: FirebaseApp): Promise<string> {
   );
 
   if (registrationPromise) {
-    // Suppress registration errors as they are not a problem for getId.
-    registrationPromise.catch(() => {});
-  }
-
-  if (installationEntry.registrationStatus === RequestStatus.COMPLETED) {
+    registrationPromise.catch(console.error);
+  } else {
     // If the installation is already registered, update the authentication
-    // token if needed. Suppress errors as they are not relevant to getId.
-    refreshAuthToken(appConfig).catch(() => {});
+    // token if needed.
+    refreshAuthToken(appConfig).catch(console.error);
   }
 
   return installationEntry.fid;